Grokie/raspberry-pi-client/requirements.txt
livekit
python-dotenv
pyaudio
requests
numpy


